requirements.txt
hrml
numpy
jinja2
PyGithub
requests
opencv-python
azure-functions